+++ src/types.rs
@@ -76,7 +76,7 @@ pub struct PretrainedModel(pub Vec<u8>);
 )]
 #[repr(u8)]
 pub enum EventCategory {
-    Unknown = 0,
+    Unspecified = 0,
     Reconnaissance = 1,
     InitialAccess,
     Execution,